% tinyrox says don't edit this manually, but it can't stop you!
\name{slice_lastdim}
\alias{slice_lastdim}
\title{Slice a contiguous range along the last dimension}
\usage{
slice_lastdim(x, from, to)
}
\arguments{
\item{x}{AnvlArray.}

\item{from}{Integer. 1-based inclusive start.}

\item{to}{Integer. 1-based inclusive end.}
}
\description{
Slice a contiguous range along the last dimension
}
