% tinyrox says don't edit this manually, but it can't stop you!
\name{yq_st_close}
\alias{yq_st_close}
\title{Close a safetensors handle}
\usage{
yq_st_close(st)
}
\arguments{
\item{st}{A handle from \code{\link{yq_st_open}} /
\code{\link{yq_st_open_sharded}}.}
}
\description{
Close a safetensors handle
}
